requirements.txt
streamlit==1.41.1
pandas==2.2.3
numpy==2.1.3
matplotlib==3.9.2
openai==1.58.1
python-dotenv==1.0.1
openpyxl==3.1.5

